account drawal add, and lack of balance exception

+++ src/main/java/com/gon/coin/demotradingcoin/InitDb.java
@@ -47,7 +47,7 @@ public class InitDb {
             em.persist(coin);
         }
         public void dbInit3(){
-            Account account1 = createAccount("국민", "123456", 1000);
+            Account account1 = createAccount("국민", "123456", Double.valueOf(1000));
             em.persist(account1);
             MemberDto memberA=createMember("memberA","1234");
             memberService.signUp(memberA);
@@ -55,7 +55,7 @@ public class InitDb {
             member.setAccount(account1);
         }
         public void dbInit4(){
-            Account account1 = createAccount("우리", "703456", 2000);
+            Account account1 = createAccount("우리", "703456", Double.valueOf(2000));
             em.persist(account1);
             MemberDto memberB=createMember("memberB","1234");
             memberService.signUp(memberB);
@@ -68,7 +68,7 @@ public class InitDb {
             MemberDto memberDto = new MemberDto(username,password);
             return memberDto;
         }
-        private Account createAccount(String bankName,String bankCode,int SumOfMoney){
+        private Account createAccount(String bankName,String bankCode,Double SumOfMoney){
             Account account= new Account(bankName,bankCode,SumOfMoney);
 
             return account;
